# exec: handle NULL group by keys in hash aggregator