--- a/pkg/sql/exec/hash_aggregator.go
+++ b/pkg/sql/exec/hash_aggregator.go
@@ -110,6 +110,7 @@ func NewHashAggregator(
 		colTypes,
 		groupCols,
 		outCols,
+		true, /* allowNullEquality */
 	)
 
 	builder := makeHashJoinBuilder(