@rafiss 2019-07-17
benchmark numbers aren't affected too much

```
name                                                                   old time/op    new time/op    delta
HashJoiner/nulls=false/fullOuter=false/distinct=true/rows=2048-24         562µs ± 4%     567µs ± 7%    ~     (p=0.436 n=10+10)
HashJoiner/nulls=false/fullOuter=false/distinct=true/rows=262144-24      44.0ms ± 2%    43.0ms ± 2%  -2.30%  (p=0.000 n=10+10)
HashJoiner/nulls=false/fullOuter=false/distinct=true/rows=4194304-24      668ms ± 1%     661ms ± 2%  -1.16%  (p=0.002 n=8+9)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=2048-24        545µs ± 1%     548µs ± 2%    ~     (p=0.173 n=8+10)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=262144-24     43.9ms ± 2%    42.7ms ± 4%  -2.59%  (p=0.002 n=10+10)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=4194304-24     1.21s ± 7%     1.15s ± 4%  -5.29%  (p=0.034 n=10+8)
HashJoiner/nulls=false/fullOuter=true/distinct=true/rows=2048-24          583µs ± 1%     561µs ± 4%  -3.71%  (p=0.000 n=9+10)
HashJoiner/nulls=false/fullOuter=true/distinct=true/rows=262144-24       45.1ms ± 1%    44.3ms ± 1%  -1.81%  (p=0.000 n=10+10)
HashJoiner/nulls=false/fullOuter=true/distinct=true/rows=4194304-24       690ms ± 2%     679ms ± 2%  -1.61%  (p=0.002 n=10+9)
HashJoiner/nulls=false/fullOuter=true/distinct=false/rows=2048-24         558µs ± 2%     530µs ± 9%  -5.07%  (p=0.001 n=10+10)
HashJoiner/nulls=false/fullOuter=true/distinct=false/rows=262144-24      43.4ms ± 3%    41.5ms ± 2%  -4.42%  (p=0.000 n=10+10)
HashJoiner/nulls=false/fullOuter=true/distinct=false/rows=4194304-24      1.00s ± 5%     1.02s ± 5%    ~     (p=0.280 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=true/rows=2048-24          643µs ± 3%     662µs ± 1%  +2.93%  (p=0.000 n=10+9)
HashJoiner/nulls=true/fullOuter=false/distinct=true/rows=262144-24       53.2ms ± 1%    54.2ms ± 1%  +2.00%  (p=0.000 n=9+10)
HashJoiner/nulls=true/fullOuter=false/distinct=true/rows=4194304-24       811ms ± 1%     827ms ± 1%  +2.01%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=false/rows=2048-24         624µs ± 4%     633µs ± 3%    ~     (p=0.190 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=false/rows=262144-24      57.2ms ± 2%    54.9ms ± 1%  -3.97%  (p=0.000 n=9+9)
HashJoiner/nulls=true/fullOuter=false/distinct=false/rows=4194304-24      1.38s ± 4%     1.39s ± 6%    ~     (p=0.529 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=true/rows=2048-24           656µs ± 1%     665µs ± 1%  +1.35%  (p=0.002 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=true/rows=262144-24        54.4ms ± 1%    56.1ms ± 3%  +3.18%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=true/rows=4194304-24        824ms ± 1%     844ms ± 1%  +2.41%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=false/rows=2048-24          625µs ± 2%     646µs ± 1%  +3.32%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=false/rows=262144-24       53.6ms ± 2%    53.0ms ± 1%  -1.09%  (p=0.043 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=false/rows=4194304-24       1.14s ± 3%     1.17s ± 3%    ~     (p=0.089 n=10+10)

name                                                                   old speed      new speed      delta
HashJoiner/nulls=false/fullOuter=false/distinct=true/rows=2048-24       233MB/s ± 4%   232MB/s ± 7%    ~     (p=0.436 n=10+10)
HashJoiner/nulls=false/fullOuter=false/distinct=true/rows=262144-24     381MB/s ± 2%   390MB/s ± 2%  +2.35%  (p=0.000 n=10+10)
HashJoiner/nulls=false/fullOuter=false/distinct=true/rows=4194304-24    402MB/s ± 1%   406MB/s ± 2%  +1.18%  (p=0.002 n=8+9)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=2048-24      240MB/s ± 1%   239MB/s ± 2%    ~     (p=0.173 n=8+10)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=262144-24    382MB/s ± 2%   393MB/s ± 3%  +2.69%  (p=0.002 n=10+10)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=4194304-24   221MB/s ± 8%   233MB/s ± 4%  +5.42%  (p=0.034 n=10+8)
HashJoiner/nulls=false/fullOuter=true/distinct=true/rows=2048-24        225MB/s ± 1%   234MB/s ± 4%  +3.91%  (p=0.000 n=9+10)
HashJoiner/nulls=false/fullOuter=true/distinct=true/rows=262144-24      372MB/s ± 1%   379MB/s ± 1%  +1.84%  (p=0.000 n=10+10)
HashJoiner/nulls=false/fullOuter=true/distinct=true/rows=4194304-24     389MB/s ± 2%   396MB/s ± 2%  +1.63%  (p=0.002 n=10+9)
HashJoiner/nulls=false/fullOuter=true/distinct=false/rows=2048-24       235MB/s ± 2%   248MB/s ± 9%  +5.51%  (p=0.001 n=10+10)
HashJoiner/nulls=false/fullOuter=true/distinct=false/rows=262144-24     387MB/s ± 3%   405MB/s ± 2%  +4.61%  (p=0.000 n=10+10)
HashJoiner/nulls=false/fullOuter=true/distinct=false/rows=4194304-24    269MB/s ± 5%   264MB/s ± 5%    ~     (p=0.280 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=true/rows=2048-24        204MB/s ± 3%   198MB/s ± 1%  -2.86%  (p=0.000 n=10+9)
HashJoiner/nulls=true/fullOuter=false/distinct=true/rows=262144-24      316MB/s ± 1%   309MB/s ± 1%  -1.96%  (p=0.000 n=9+10)
HashJoiner/nulls=true/fullOuter=false/distinct=true/rows=4194304-24     331MB/s ± 1%   324MB/s ± 1%  -1.97%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=false/rows=2048-24       210MB/s ± 4%   207MB/s ± 3%    ~     (p=0.190 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=false/rows=262144-24     293MB/s ± 2%   306MB/s ± 1%  +4.13%  (p=0.000 n=9+9)
HashJoiner/nulls=true/fullOuter=false/distinct=false/rows=4194304-24    195MB/s ± 4%   193MB/s ± 6%    ~     (p=0.529 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=true/rows=2048-24         200MB/s ± 1%   197MB/s ± 1%  -1.34%  (p=0.002 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=true/rows=262144-24       309MB/s ± 1%   299MB/s ± 3%  -3.07%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=true/rows=4194304-24      326MB/s ± 1%   318MB/s ± 1%  -2.36%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=false/rows=2048-24        210MB/s ± 2%   203MB/s ± 1%  -3.22%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=false/rows=262144-24      313MB/s ± 2%   316MB/s ± 1%  +1.10%  (p=0.043 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=false/rows=4194304-24     235MB/s ± 3%   230MB/s ± 3%    ~     (p=0.089 n=10+10)

name                                                                   old alloc/op   new alloc/op   delta
HashJoiner/nulls=false/fullOuter=false/distinct=true/rows=2048-24         805kB ± 0%     805kB ± 0%  +0.00%  (p=0.000 n=9+8)
HashJoiner/nulls=false/fullOuter=false/distinct=true/rows=262144-24      46.4MB ± 0%    46.4MB ± 0%  +0.00%  (p=0.000 n=9+10)
HashJoiner/nulls=false/fullOuter=false/distinct=true/rows=4194304-24      697MB ± 0%     697MB ± 0%  +0.00%  (p=0.001 n=8+10)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=2048-24        809kB ± 0%     809kB ± 0%  +0.00%  (p=0.000 n=10+8)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=262144-24     44.6MB ± 0%    44.6MB ± 0%  +0.00%  (p=0.001 n=10+9)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=4194304-24     668MB ± 0%     668MB ± 0%    ~     (p=0.222 n=10+10)
HashJoiner/nulls=false/fullOuter=true/distinct=true/rows=2048-24          808kB ± 0%     808kB ± 0%  +0.00%  (p=0.000 n=10+9)
HashJoiner/nulls=false/fullOuter=true/distinct=true/rows=262144-24       46.7MB ± 0%    46.7MB ± 0%  +0.00%  (p=0.003 n=10+10)
HashJoiner/nulls=false/fullOuter=true/distinct=true/rows=4194304-24       701MB ± 0%     701MB ± 0%  +0.00%  (p=0.000 n=9+10)
HashJoiner/nulls=false/fullOuter=true/distinct=false/rows=2048-24         812kB ± 0%     812kB ± 0%  +0.00%  (p=0.000 n=10+7)
HashJoiner/nulls=false/fullOuter=true/distinct=false/rows=262144-24      44.9MB ± 0%    44.9MB ± 0%  +0.00%  (p=0.000 n=10+10)
HashJoiner/nulls=false/fullOuter=true/distinct=false/rows=4194304-24      672MB ± 0%     672MB ± 0%    ~     (p=0.338 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=true/rows=2048-24          805kB ± 0%     805kB ± 0%  -0.00%  (p=0.000 n=10+9)
HashJoiner/nulls=true/fullOuter=false/distinct=true/rows=262144-24       46.4MB ± 0%    46.4MB ± 0%  -0.02%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=true/rows=4194304-24       697MB ± 0%     697MB ± 0%  -0.02%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=false/rows=2048-24         809kB ± 0%     809kB ± 0%  -0.00%  (p=0.000 n=9+10)
HashJoiner/nulls=true/fullOuter=false/distinct=false/rows=262144-24      44.6MB ± 0%    44.6MB ± 0%  -0.01%  (p=0.000 n=9+9)
HashJoiner/nulls=true/fullOuter=false/distinct=false/rows=4194304-24      668MB ± 0%     668MB ± 0%  -0.01%  (p=0.000 n=6+10)
HashJoiner/nulls=true/fullOuter=true/distinct=true/rows=2048-24           808kB ± 0%     808kB ± 0%  -0.00%  (p=0.000 n=9+8)
HashJoiner/nulls=true/fullOuter=true/distinct=true/rows=262144-24        46.7MB ± 0%    46.7MB ± 0%  -0.02%  (p=0.000 n=8+10)
HashJoiner/nulls=true/fullOuter=true/distinct=true/rows=4194304-24        701MB ± 0%     701MB ± 0%  -0.02%  (p=0.000 n=10+8)
HashJoiner/nulls=true/fullOuter=true/distinct=false/rows=2048-24          812kB ± 0%     812kB ± 0%  -0.00%  (p=0.000 n=9+6)
HashJoiner/nulls=true/fullOuter=true/distinct=false/rows=262144-24       44.9MB ± 0%    44.9MB ± 0%  -0.01%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=false/rows=4194304-24       672MB ± 0%     672MB ± 0%  -0.01%  (p=0.000 n=9+10)

name                                                                   old allocs/op  new allocs/op  delta
HashJoiner/nulls=false/fullOuter=false/distinct=true/rows=2048-24         8.28k ± 0%     8.28k ± 0%    ~     (all equal)
HashJoiner/nulls=false/fullOuter=false/distinct=true/rows=262144-24       1.05M ± 0%     1.05M ± 0%    ~     (all equal)
HashJoiner/nulls=false/fullOuter=false/distinct=true/rows=4194304-24      16.8M ± 0%     16.8M ± 0%    ~     (all equal)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=2048-24        6.24k ± 0%     6.24k ± 0%    ~     (all equal)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=262144-24       527k ± 0%      527k ± 0%    ~     (all equal)
HashJoiner/nulls=false/fullOuter=false/distinct=false/rows=4194304-24     8.40M ± 0%     8.40M ± 0%    ~     (p=0.721 n=10+10)
HashJoiner/nulls=false/fullOuter=true/distinct=true/rows=2048-24          8.28k ± 0%     8.28k ± 0%    ~     (all equal)
HashJoiner/nulls=false/fullOuter=true/distinct=true/rows=262144-24        1.05M ± 0%     1.05M ± 0%    ~     (all equal)
HashJoiner/nulls=false/fullOuter=true/distinct=true/rows=4194304-24       16.8M ± 0%     16.8M ± 0%    ~     (p=0.248 n=9+10)
HashJoiner/nulls=false/fullOuter=true/distinct=false/rows=2048-24         6.24k ± 0%     6.24k ± 0%    ~     (all equal)
HashJoiner/nulls=false/fullOuter=true/distinct=false/rows=262144-24        527k ± 0%      527k ± 0%    ~     (all equal)
HashJoiner/nulls=false/fullOuter=true/distinct=false/rows=4194304-24      8.40M ± 0%     8.40M ± 0%    ~     (p=0.173 n=10+9)
HashJoiner/nulls=true/fullOuter=false/distinct=true/rows=2048-24          8.28k ± 0%     8.27k ± 0%  -0.10%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=true/rows=262144-24        1.05M ± 0%     1.05M ± 0%  -0.10%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=true/rows=4194304-24       16.8M ± 0%     16.8M ± 0%  -0.10%  (p=0.001 n=8+9)
HashJoiner/nulls=true/fullOuter=false/distinct=false/rows=2048-24         6.24k ± 0%     6.23k ± 0%  -0.10%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=false/rows=262144-24        527k ± 0%      527k ± 0%  -0.10%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=false/distinct=false/rows=4194304-24      8.41M ± 0%     8.40M ± 0%  -0.10%  (p=0.000 n=7+10)
HashJoiner/nulls=true/fullOuter=true/distinct=true/rows=2048-24           8.28k ± 0%     8.28k ± 0%  -0.10%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=true/rows=262144-24         1.05M ± 0%     1.05M ± 0%  -0.10%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=true/rows=4194304-24        16.8M ± 0%     16.8M ± 0%  -0.10%  (p=0.002 n=8+10)
HashJoiner/nulls=true/fullOuter=true/distinct=false/rows=2048-24          6.24k ± 0%     6.23k ± 0%  -0.10%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=false/rows=262144-24         527k ± 0%      527k ± 0%  -0.10%  (p=0.000 n=10+10)
HashJoiner/nulls=true/fullOuter=true/distinct=false/rows=4194304-24       8.40M ± 0%     8.40M ± 0%  -0.10%  (p=0.000 n=9+10)
```